+++ sdgym/_run_benchmark/upload_benchmark_results.py
@@ -6,7 +6,7 @@ from datetime import datetime
 import boto3
 
 from sdgym.result_writer import S3ResultsWriter
-from sdgym.run_benchmark import OUTPUT_DESTINATION_AWS, RESULT_UPLOADED
+from sdgym._run_benchmark import OUTPUT_DESTINATION_AWS, RESULTS_UPLOADED
 from sdgym.sdgym_result_explorer.result_explorer import SDGymResultsExplorer
 
 
@@ -25,7 +25,7 @@ def get_run_name(date_str):
 
 
 def main():
-    if RESULT_UPLOADED:
+    if RESULTS_UPLOADED:
         print('Benchmark results have already been uploaded. Exiting.')
         sys.exit(0)
