+++ tasks.py
@@ -206,4 +206,10 @@ def rmdir(c, path):
 @task
 def sdgym_benchmark(c):
     """Run the SDGym benchmark."""
-    c.run('python sdgym/_run_benchmark.py')
+    c.run('python sdgym/_run_benchmark/run_benchmark.py')
+
+@task(help={"date": "Benchmark date in YYYY-MM-DD format (default: today with day=01)"})
+def upload_benchmark_results(c, date=None):
+    """Upload the benchmark results to S3."""
+    date_arg = f"--date {date}" if date else ""
+    c.run(f'python sdgym/_run_benchmark/upload_benchmark_results.py {date_arg}')
